Elasticache (#2)
* Allow service broker to use IAM proxy, ie kube2iam.

Helm 3 validates kubernetes objects. Therefore the helm chart will fail
to validate if the accesskey and secretkey are not passed.

* Change version for DC.

* Customize Elasticache for Document Cloud

* Add function push to makefile.

* Get Subnets also returns vpc id.

Add layer with cfnresponse and requests.

Update README.md

Remove small plan.

The single node configuration for mobile-sig-capture, on which the small
plan is based is incorrect.  Removing the small plan and changing the
production plan to match what mobile-sig-capture should be.